se acomodo barrera de seguridad

+++ controllers/admin.controller.php
@@ -7,70 +7,95 @@ require_once 'views/admin.view.php';
 require_once 'views/public.view.php';
 require_once 'helpers/auth.helper.php';
 
-class AdminController{
+class AdminController
+{
 
     private $modelBandas;
     private $modelCanciones;
     private $modelLogin;
     private $viewAdmin;
     private $viewPublic;
+    private $helper;
 
     public function __construct()
     {
-        authHelper::checkLogged();
         $this->modelBandas = new BandasModel();
         $this->modelCanciones = new CancionesModel();
         $this->modelLogin = new LoginModel();
         $this->viewAdmin = new AdminView();
         $this->viewPublic = new PublicView();
+        $this->helper = new AuthHelper();
     }
 
     //muestra un formulario vacio para cargar una banda
-    public function formBanda(){
-        $this->viewAdmin->formBandaAdd();
+    public function formBanda()
+    {
+        if ($this->helper->checkLogged()) {
+            $this->viewAdmin->formBandaAdd();
+        } else {
+            header('Location: ' . BASE_URL . 'suscribirse');
+        }
     }
 
     //guarda una nueva banda
-    public function addBanda(){
-        if(empty($_POST['nombre'])){
-            $this->viewAdmin->showError("No completo los datos obligatorios");
-        } else{
-            $banda = $this->modelBandas->getName($_POST['nombre']);
-            if(!empty($banda)){
-                $this->viewAdmin->showError("La banda ya existe");
-            } else{
-                $this->modelBandas->insert($_POST['nombre']);
-                header('Location: ' . BASE_URL . 'listaBandas');
+    public function addBanda()
+    {
+        if ($this->helper->checkLogged()) {
+            if (empty($_POST['nombre'])) {
+                $this->viewAdmin->showError("No completo los datos obligatorios");
+            } else {
+                $banda = $this->modelBandas->getName($_POST['nombre']);
+                if (!empty($banda)) {
+                    $this->viewAdmin->showError("La banda ya existe");
+                } else {
+                    $this->modelBandas->insert($_POST['nombre']);
+                    header('Location: ' . BASE_URL . 'listaBandas');
+                }
             }
+        } else {
+            header('Location: ' . BASE_URL . 'suscribirse');
         }
     }
 
     //muestra formulario para editar una banda
     public function editBanda($id_banda)
     {
-        $banda = $this->modelBandas->get($id_banda);
-        $this->viewAdmin->showFormEditBanda($banda);
+        if ($this->helper->checkLogged()) {
+            $banda = $this->modelBandas->get($id_banda);
+            $this->viewAdmin->showFormEditBanda($banda);
+        } else {
+            header('Location: ' . BASE_URL . 'suscribirse');
+        }
     }
 
     //modifica una banda
     public function modifyBanda()
     {
-        if (empty($_POST['nombre'])) {
-            $banda = $this->modelBandas->getName($_POST['nombre']);
-            $this->viewAdmin->showFormEditBanda($banda, "completar todos los campos");
-        }
-        else{
-            $this->modelBandas->update($_POST['nombre'], $_POST['id']);
-            $banda = $this->modelBandas->getName($_POST['nombre']);
-            $this->viewAdmin->showFormEditBanda($banda, "los cambios se guardaron correctamente");
+        if ($this->helper->checkLogged()) {
+
+            if (empty($_POST['nombre'])) {
+                $banda = $this->modelBandas->getName($_POST['nombre']);
+                $this->viewAdmin->showFormEditBanda($banda, "completar todos los campos");
+            } else {
+                $this->modelBandas->update($_POST['nombre'], $_POST['id']);
+                $banda = $this->modelBandas->getName($_POST['nombre']);
+                $this->viewAdmin->showFormEditBanda($banda, "los cambios se guardaron correctamente");
+                header('Location: ' . BASE_URL . 'listaBandas');
+            }
+        } else {
+            header('Location: ' . BASE_URL . 'suscribirse');
         }
     }
 
     //elimina una banda
     public function deleteBanda($id_Banda)
     {
-        $this->modelBandas->delete($id_Banda);
-        header('Location: ' . BASE_URL . 'listaBandas');
+        if ($this->helper->checkLogged()) {
+            $this->modelBandas->delete($id_Banda);
+            header('Location: ' . BASE_URL . 'listaBandas');
+        } else {
+            header('Location: ' . BASE_URL . 'suscribirse');
+        }
     }
 
     public function showError($msg)
@@ -78,71 +103,91 @@ class AdminController{
         $this->viewAdmin->showError($msg);
     }
 
-    public function formCancion(){
-        $bandas=$this->modelBandas->getAll();
-        $this->viewAdmin->formCancionAdd($bandas);
+    public function formCancion()
+    {
+        if ($this->helper->checkLogged()) {
+            $bandas = $this->modelBandas->getAll();
+            $this->viewAdmin->formCancionAdd($bandas);
+        } else {
+            header('Location: ' . BASE_URL . 'suscribirse');
+        }
     }
 
-    public function addCancion(){
-        $nombre_cancion = $_POST["nombreCancion"];
-        $id_banda = $_POST["idBanda"];
-        $letra_cancion = $_POST["letraCancion"];
-        $genero_cancion = $_POST["genero"];
+    public function addCancion()
+    {
+        if ($this->helper->checkLogged()) {
+            $nombre_cancion = $_POST["nombreCancion"];
+            $id_banda = $_POST["idBanda"];
+            $letra_cancion = $_POST["letraCancion"];
+            $genero_cancion = $_POST["genero"];
 
-        if(!empty($nombre_cancion)&&!empty($id_banda)&&!empty($letra_cancion)&&!empty($genero_cancion)){
-            
-           $agregada = $this->modelCanciones->addCancion($nombre_cancion,$letra_cancion,$genero_cancion,$id_banda);
-            if($agregada){
+            if (!empty($nombre_cancion) && !empty($id_banda) && !empty($letra_cancion) && !empty($genero_cancion)) {
+
+                $agregada = $this->modelCanciones->addCancion($nombre_cancion, $letra_cancion, $genero_cancion, $id_banda);
+                if ($agregada) {
+
+                    header('Location: ' . BASE_URL . 'listaCanciones');
+                } else {
+
+                    $this->showError("ERROR! no se pudo agregar la canción, intente nuevamente");
+                }
+            } else {
+                $this->showError("ERROR! quedaron campos vacios");
+            }
+        } else {
+            header('Location: ' . BASE_URL . 'suscribirse');
+        }
+    }
+
+    public function formEditCancion($id_cancion)
+    {
+        if ($this->helper->checkLogged()) {
+            $cancion = $this->modelCanciones->cancion($id_cancion);
+            $bandas = $this->modelBandas->getAll();
+            $this->viewAdmin->formCancionEdit($cancion, $bandas);
+        } else {
+            header('Location: ' . BASE_URL . 'suscribirse');
+        }
+    }
+
+    public function editCancion()
+    {
+        if ($this->helper->checkLogged()) {
+            $nombre_cancion = $_POST["nombreCancion"];
+            $id_banda = $_POST["idBanda"];
+            $letra_cancion = $_POST["letraCancion"];
+            $genero_cancion = $_POST["genero"];
+            $id_cancion = $_POST["id_cancion"];
+            if (!empty($nombre_cancion) && !empty($id_banda) && !empty($letra_cancion) && !empty($genero_cancion) && !empty($id_cancion)) {
+
+                $editada = $this->modelCanciones->editCancion($nombre_cancion, $letra_cancion, $genero_cancion, $id_banda, $id_cancion);
+                if ($editada) {
+
+                    header('Location: ' . BASE_URL . 'listaCanciones');
+                } else {
+
+                    $this->showError("ERROR! no se pudo editar la canción, intente nuevamente");
+                }
+            } else {
+                $this->showError("ERROR! quedaron campos vacios");
+            }
+        } else {
+            header('Location: ' . BASE_URL . 'suscribirse');
+        }
+    }
+    public function deleteCancion($id_cancion)
+    {
+        if ($this->helper->checkLogged()) {
+            $eliminada = $this->modelCanciones->delete($id_cancion);
+            if ($eliminada) {
 
                 header('Location: ' . BASE_URL . 'listaCanciones');
-            }else{
+            } else {
 
-                $this->showError("ERROR! no se pudo agregar la canción, intente nuevamente");
+                $this->showError("ERROR! no se pudo eliminar la canción, intente nuevamente");
             }
-        }else{
-            $this->showError("ERROR! quedaron campos vacios");
-        }
-      
-    }
-    public function formEditCancion($id_cancion){
-
-        $cancion = $this->modelCanciones->cancion($id_cancion);
-        $bandas=$this->modelBandas->getAll();
-        $this->viewAdmin->formCancionEdit($cancion, $bandas);
-    }
-
-    public function editCancion(){
-
-        $nombre_cancion = $_POST["nombreCancion"];
-        $id_banda = $_POST["idBanda"];
-        $letra_cancion = $_POST["letraCancion"];
-        $genero_cancion = $_POST["genero"];
-        $id_cancion=$_POST["id_cancion"];
-        if(!empty($nombre_cancion)&&!empty($id_banda)&&!empty($letra_cancion)&&!empty($genero_cancion)&&!empty($id_cancion)){
-            
-           $editada = $this->modelCanciones->editCancion($nombre_cancion,$letra_cancion,$genero_cancion,$id_banda,$id_cancion);
-            if($editada){
-
-                header('Location: ' . BASE_URL . 'listaCanciones');
-            }else{
-
-                $this->showError("ERROR! no se pudo editar la canción, intente nuevamente");
-            }
-        }else{
-            $this->showError("ERROR! quedaron campos vacios");
-        }
-
-       
-    }
-    public function deleteCancion($id_cancion){
-        $eliminada = $this->modelCanciones->delete($id_cancion);
-        if($eliminada){
-
-            header('Location: ' . BASE_URL . 'listaCanciones');
-        }else{
-
-            $this->showError("ERROR! no se pudo eliminar la canción, intente nuevamente");
+        } else {
+            header('Location: ' . BASE_URL . 'suscribirse');
         }
     }
-      
 }
